pyproject.toml
[tool.poetry]
name = "scrapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests2 = "^2.16.0"
beautifulsoup4 = "^4.9.3"
flask = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
